backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
requests==2.32.5
firebase-admin==7.1.0
python-dotenv==1.0.1
loguru==0.7.3
pydantic==2.11.9
passlib[bcrypt]==1.7.4
google-api-python-client==2.159.0
google-auth==2.38.0